combined ND analysis complete for now

+++ dimensional_analysis_functions.py
@@ -1,6 +1,7 @@
 import os
 import numpy as np
 import matplotlib.pyplot as plt
+from fractions import Fraction
 from matplotlib.lines import Line2D
 
 def plot_rig_exponents(color_opt, time, it, fig_folder, w_rms, b_center, bw, rp, T, S, z_nd, zf_nd, Ri_g, case_names, exponents = [-0.5, -1/3, -0.25, 0.0, 0.25, 1/3, 0.5]):
@@ -34,7 +35,7 @@ def plot_rig_exponents(color_opt, time, it, fig_folder, w_rms, b_center, bw, rp,
                     zf_nd[:, i], color=color_opt[i])
         ax.ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
         ax.set_title(f'Ri_g^{exp:.2f}', fontsize = 16)
-        ax.set_xlabel(r'$w_{{rms}}/\sqrt{\text{g l}_{j}} \cdot Ri_g^{exp}$', fontsize = 16)
+        ax.set_xlabel(r"$w_{{rms}}/\sqrt{\text{g l}_{j}} \cdot Ri_g^{exp}$", fontsize = 16)
     #axes[0, 0].legend()
 
     for ax, exp in zip(axes[1, :], exponents):
@@ -44,7 +45,7 @@ def plot_rig_exponents(color_opt, time, it, fig_folder, w_rms, b_center, bw, rp,
                     z_nd[:, i], color=color_opt[i])
         ax.ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
         ax.set_title(f'Ri_g^{exp:.2f}', fontsize = 16)
-        ax.set_xlabel(r'$b_{centerline}/g \cdot Ri_g^{exp}$', fontsize = 16)
+        ax.set_xlabel(r"$b_{centerline}/g \cdot Ri_g^{exp}$", fontsize = 16)
     #axes[1, 0].legend()
 
     for ax, exp in zip(axes[2, :], exponents):
@@ -129,7 +130,7 @@ def plot_Fr_exponents(color_opt, time, it, fig_folder, w_rms, b_center, bw, rp, 
                     zf_nd[:, i], color=color_opt[i])
         ax.ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
         ax.set_title(f'Fr^{exp:.2f}', fontsize = 16)
-        ax.set_xlabel(r'$w_{{rms}}/\sqrt{\text{g l}_{j}} \cdot Fr^{exp}$', fontsize = 16)
+        ax.set_xlabel(r"$w_{{rms}}/\sqrt{\text{g l}_{j}} \cdot Fr^{exp}$", fontsize = 16)
     #axes[0, 0].legend()
 
     for ax, exp in zip(axes[1, :], exponents):
@@ -139,7 +140,7 @@ def plot_Fr_exponents(color_opt, time, it, fig_folder, w_rms, b_center, bw, rp, 
                     z_nd[:, i], color=color_opt[i])
         ax.ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
         ax.set_title(f'Fr^{exp:.2f}', fontsize = 16)
-        ax.set_xlabel(r'$b_{centerline}/g \cdot Fr^{exp}$', fontsize = 16)
+        ax.set_xlabel(r"$b_{centerline}/g \cdot Fr^{exp}$", fontsize = 16)
     #axes[1, 0].legend()
 
     for ax, exp in zip(axes[2, :], exponents):
@@ -222,7 +223,7 @@ def plot_mld_exponents(color_opt, time, it, fig_folder, w_rms, b_center, bw, rp,
                     zf_nd[:, i], color=color_opt[i])
         ax.ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
         ax.set_title(f'mld^{exp:.2f}', fontsize = 16)
-        ax.set_xlabel(r'$w_{{rms}}/\sqrt{\text{g l}_{j}} \cdot mld^{exp}$', fontsize = 16)
+        ax.set_xlabel(r"$w_{{rms}}/\sqrt{\text{g l}_{j}} \cdot mld^{exp}$", fontsize = 16)
     #axes[0, 0].legend()
 
     for ax, exp in zip(axes[1, :], exponents):
@@ -232,7 +233,7 @@ def plot_mld_exponents(color_opt, time, it, fig_folder, w_rms, b_center, bw, rp,
                     z_nd[:, i], color=color_opt[i])
         ax.ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
         ax.set_title(f'mld^{exp:.2f}', fontsize = 16)
-        ax.set_xlabel(r'$b_{centerline}/g \cdot mld^{exp}$', fontsize = 16)
+        ax.set_xlabel(r"$b_{centerline}/g \cdot mld^{exp}$", fontsize = 16)
     #axes[1, 0].legend()
 
     for ax, exp in zip(axes[2, :], exponents):
@@ -286,14 +287,16 @@ def plot_mld_exponents(color_opt, time, it, fig_folder, w_rms, b_center, bw, rp,
     
 
 def plot_combo_exponents(color_opt, time, it, fig_folder, w_rms, b_center, bw, rp, T, S, z_nd, zf_nd, vars_exps, Ri_g, Fr, mld, case_names):
-    vars_str = vars_exps
+    NDs = [rf"MLD$^", rf"Ri_g$^", rf"Fr$^"] 
+    NDs_filtered = [[("" if str(Fraction(x).limit_denominator()) == '0' 
+                else NDs[j] + "{"+str(Fraction(x).limit_denominator())+"}$")
+                for j, x in enumerate(row)] for row in vars_exps]
+    vars_str = [''.join(row) for row in NDs_filtered]
     num_cases = len(case_names)
-    scale = np.ones(7) 
-    scale[-1] = 0.1
-    gridspec_kw={'height_ratios': scale}
-    fig, axes = plt.subplots(1, len(scale), figsize=(25, 7), sharey=True, gridspec_kw = gridspec_kw)
+    gridspec_kw={'height_ratios': [1, 0.3]}
+    fig, axes = plt.subplots(2, len(NDs_filtered), figsize=(25, 7), sharey=True, gridspec_kw = gridspec_kw)
     plt.subplots_adjust(top=0.9)
-    for a in axes[-1]:
+    for a in axes[-1, :]:
         a.remove()
     case_handles = [Line2D([0], [0], color=color_opt[i], linestyle='solid', label=case_names[i])for i in range(num_cases)]
     fig.legend(handles=case_handles,
@@ -311,47 +314,54 @@ def plot_combo_exponents(color_opt, time, it, fig_folder, w_rms, b_center, bw, r
     axes[5] = ND average salinity vs z_nd varied exponent of all
     """
     for i in range(num_cases):
-        axes[0].plot(w_rms[:, i] * mld[0]**vars_exps[0, 0] * Ri_g[0]**vars_exps[0, 1] * Fr[0]**vars_exps[0, 2], 
+        axes[0, 0].plot(w_rms[:, i] * mld[0]**vars_exps[0, 0] * Ri_g[0]**vars_exps[0, 1] * Fr[0]**vars_exps[0, 2], 
                 zf_nd[:, i], color=color_opt[i])
-    axes[0].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
-    axes[0].set_xlabel(r'$w_{{rms}}/\sqrt{\text{g l}_{j}} \cdot mld^{exp}$', fontsize = 16)
+    axes[0, 0].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
+    axes[0, 0].set_xlabel(rf"$w_{{rms}}/\sqrt{{\text{{g l}}_{{j}}}}\cdot$ {vars_str[0]}", fontsize = 16)
 
     for i in range(num_cases):
-        axes[1].plot(b_center[:, i] * mld[1]**vars_exps[1, 0] * Ri_g[1]**vars_exps[1, 1] * Fr[1]**vars_exps[1, 2], 
+        axes[0, 1].plot(b_center[:, i] * mld[1]**vars_exps[1, 0] * Ri_g[1]**vars_exps[1, 1] * Fr[1]**vars_exps[1, 2], 
                 z_nd[:, i], color=color_opt[i])
-    axes[1].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
-    axes[1].set_xlabel(r'$b_{centerline}/g \cdot mld^{exp}$', fontsize = 16)
+    axes[0, 1].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
+    axes[0, 1].set_xlabel(rf"$b_{{centerline}}/g \cdot$ {vars_str[1]}", fontsize = 16)
 
     for i in range(num_cases):
-        axes[2].plot(bw[:, i] * mld[2]**vars_exps[2, 0] * Ri_g[2]**vars_exps[2, 1] * Fr[2]**vars_exps[2, 2], 
+        axes[0, 2].plot(bw[:, i] * mld[2]**vars_exps[2, 0] * Ri_g[2]**vars_exps[2, 1] * Fr[2]**vars_exps[2, 2], 
                 z_nd[:, i], color=color_opt[i])
-    axes[2].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
-    axes[2].set_xlabel(r"$(\langle b'w'\rangle_{xy}/\sqrt{\text{g}^3 \text{r}_{j}})\cdot mld^{exp}$", fontsize = 16)
+    axes[0, 2].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
+    axes[0, 2].set_xlabel(rf"$(\langle b'w'\rangle_{{xy}}/\sqrt{{\text{{g}}^3 \text{{r}}_{{j}}}})\cdot$ {vars_str[2]}", fontsize = 16)
 
     for i in range(num_cases):
-        axes[3].plot(rp[:, i] * mld[3]**vars_exps[3, 0] * Ri_g[3]**vars_exps[3, 1] * Fr[3]**vars_exps[3, 2], 
+        axes[0, 3].plot(rp[:, i] * mld[3]**vars_exps[3, 0] * Ri_g[3]**vars_exps[3, 1] * Fr[3]**vars_exps[3, 2], 
                 z_nd[:, i], color=color_opt[i])
-    axes[3].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
-    axes[3].set_xlabel(r"(r/l$_{j})\cdot mld^{exp}$", fontsize = 16)
+    axes[0, 3].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
+    axes[0, 3].set_xlabel(rf"(r/l$_{{j}})\cdot$ {vars_str[3]}", fontsize = 16)
 
     for i in range(num_cases):
-        axes[4].plot(T[:, i] * mld[4]**vars_exps[4, 0] * Ri_g[4]**vars_exps[4, 1] * Fr[4]**vars_exps[4, 2], 
+        axes[0, 4].plot(T[:, i] * mld[4]**vars_exps[4, 0] * Ri_g[4]**vars_exps[4, 1] * Fr[4]**vars_exps[4, 2],
                 z_nd[:, i], color=color_opt[i])
-    axes[4].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
-    axes[4].set_xlabel(r"$(\text{T'}_{\text{centerline}}\alpha)\cdot mld^{exp}$", fontsize = 16)
+    axes[0, 4].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
+    axes[0, 4].set_xlabel(rf"$(\text{{T'}}_{{\text{{centerline}}}}\alpha)\cdot$ {vars_str[4]}", fontsize = 16)
 
     for i in range(num_cases):
-        axes[5].plot(S[:, i] * mld[5]**vars_exps[5, 0] * Ri_g[5]**vars_exps[5, 1] * Fr[5]**vars_exps[5, 2], 
+        axes[0, 5].plot(S[:, i] * mld[5]**vars_exps[5, 0] * Ri_g[5]**vars_exps[5, 1] * Fr[5]**vars_exps[5, 2], 
                 z_nd[:, i], color=color_opt[i])
-    axes[5].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
-    axes[5].set_xlabel(r"($\langle$C$\rangle_{\text{xy}} \beta)\cdot mld^{exp}$", fontsize = 16)
+    axes[0, 5].ticklabel_format(axis='x', style='sci', scilimits=(-3,2), useMathText=True)
+    axes[0, 5].set_xlabel(rf"($\langle$C$\rangle_{{\text{{xy}}}} \beta)\cdot$ {vars_str[5]}", fontsize = 16)
 
 
     plt.tight_layout()
 
     # --- Save Frame ---
     frame_path = os.path.join(fig_folder, f"it{it:04d}_combined.png")
-    plt.savefig(frame_path)
+    i = 0
+    while True:
+        i += 1
+        frame_path = os.path.join(fig_folder, f"it{it:04d}_combined_{i}.png")
+        if os.path.exists(frame_path):
+            continue
+        plt.savefig(frame_path)
+        break
     plt.close(fig)
     print(f"Time step {it + 1} captured")
